allow using AD FS as identity provider (issue #16585)
This mainly involves a simplified SOAP request for the SAML assertion.
The fields in the assertion must be customised on the AD, as it
doesn't even include a name by default.

Since the new code doesn't fork a JVM, it is significantly faster then
the 'auth.sh' script used for WSO2. As-is, however, the simplified
approach gave a 500 Internal Server Error with WSO2, so we'll leave it
as an exercise for the future.