Installer cleanup: prevent installer from ever deleting the database
Asking the user whether they want to delete the database is rather
obnoxious; at least in production, we should never, ever delete data.
Instead, we factor the drops out to a separate file, and delete the
prompt.

This way, we have one, safe script -- initdb.sh -- which initialises
the database, and fails if it already exists, and another script which
drops the table and then calls the other.