TIA: extract database structure from settings.py to a separate module
'settings.py' contains user-specified settings, so we shouldn't
overwrite it, yet we do want to alter the database structure on
upgrade. As such, the user configuration is a poor location for it.

While at it, I've extracted the list of tables from two of the
duplicated locations; for now, MoxRestFrontend continues to provide
its own list.