Fix searching by state
LoRA supports restricting searches by state similar to how it supports
attribute and relation restrictions. Unfortunately however, that code
was broken. The main culprit was inconsistent use if field and
attribute names. With this change, the following works as expected:

?gyldighed=Aktiv

This change removes the single use of the get_state_field() function
in db_helpers.py, so we euthanise it. The end result is that the
state/tilstand code is mostly consistent with relations and
attributes.