settings: allow specifying database port & host; enable monkey-patching
Importing by value is both bad form and untestable; if we instead just
import a module, we can simply override the value on that module,
making testing much easier.

The only downside is that it involves an extra dictionary lookup,
hence being ever so slightly slower. We should do local caching
instead should that prove an issue.