build_registration.restriction_to_registration(): use iteritems()
As-is, the function is quite odd: the flatten() function is equivalent
to dict.items(), and it also generates a list for each type of
restriction. Since 2to3 tries to retain that side effect, it'd wrap
each items() call in a further list() call, making the code even more
odd.

Instead, just use itertools.chain() of the item views, adopting Python
3 semantics in advance.