simplify get_valid_search_parameters() to rely on simple iteration
Calling keys() is both wasteful and less flexible; relying on simple
iteration for constructing the acceptable parameters works with other
types as well.

(When writing Python 2.7 code, keys() is eager and returns a list,
whereas it returns a “set-like view” in Python 3. The end result is
that when iterating, just use the `dict` itself -- it's just as
readable, it's more flexible due to working with lists, sets and so
on, and more simpler overall due to the more concise code.)