move the database connection to the context, and pool it
This ensures that we only ever create a single connection per request,
and also removes the hack for the `db.adapt()` function. Furthermore,
we now do the caching within the get_connection() method, it no longer
matters how many times we call it. Hence, I killed that test.

The motivation for this is that first of all, not having the adapt()
connection stick around should make it easier to kick the PostgreSQL
processes without also bouncing OIO REST, and second, it should aid
with running OIO REST in PyPy.* Also, it's nice and the documented way
of doing things. Finally, using a pool should give us a slight
speed-up when performing multiple requests.

By default, we have no minimum amount of connections; this is nice for
development, where we may want to start the webapp before the
database. We can increase it in production using the settings.

* I know, no-one cares about this but me…