setup.py & docs: obtain correct version from VERSION
The previous logic was flawed, and used the wrong location, given that
we have a `VERSION` file for this specific case. While at it, I fixed
the authors line in `setup.py` to just state “Magenta ApS”, and added
a test to verify agreement between `NEWS.rst` and `VERSION`.

The motiviation for doing this as a direct hotfix/post-release is that
our current documentation gets the version wrong.