db: don't use SELECT DISTINCT for paging and ordering
Grouping by object ID and aggregated brugervendtnoegle results in a
filtered search over 25k objects (with one effect) taking about 4
seconds to retrieve 20 items, and staying that way as the offset
increases.

The DISTINCT on brugervendtnoegle shouldn't matter much, as we expect
this value to remain mostly unchanged throughout the life of an
object. Furthermore, we expect most searches to be relative to a
certain time, and not apply to future and past.

As an added bonus, we now consider the effect and registration time
requested when sorting -- unlike authentication.