Hopefully fixes all the review commonts that I can agree to :-)
The threaded design is chosen as I consider this the simplest way
to achive a heatbeat that will come at regular intevals without
blocking the relay of the notifcation.

Heatbeats sent from the timeout will only come if no actual notifications
is recieved within the timeout period. I would very much like to be able
to deliver a continuous stream of heatbeats independent of notification
rate, which the current implementation deliers (a heartbeat at least
every 2 seconds + more if no messages is recieved within timeout).
If the wish to avoid threading is very strong, I will prefer to postpone the
heartbeat mechanism alltogether.

Overall, I see the task as an attempt at a minimal implementation until
actual customer wants us to expand on this, for this reason I have done
no attempts at fancy configuration of the AMQP modes.

A valid test procedure for this could would be to create an actual
lora object and see the notification on the AMQP-queue, however this
requires the full AMQP functionallity to be available on the test
host, I do not know if this is the case?