add basic support for counting objects
This is a quick-and-dirty implementation which is significantly faster in
most cases. The main downside is that it doesn't support restrictions,
yet, and also doesn't take effect and registration into effect.

Some benchmarks in a database with ~56k objects, all valid with
searches that yield ~800 objects:

- filtering for validity, an attribute and one relation:
  from ~2.7s to <0.1s
- filtering for validity, an attribute and two relations:
  from ~4.8s to ~4.5s

The extra relation happens to be largely redundant in the MO data
model, but the cause of the significant overhead is likely joining the
relations table twice. If we could get rid of that, near-instant
searches seem doable.

(I've added a few basic tests -- among other things, they confirm that
counting documents doesn't work, yet.)