Make brugervendtnoegle mandatory (NOT NULL)
According to the OIO standard version 1.1, brugervendtnøgle should always be
specified. In the database this means not null.

Previously, we had this specified in `oio_rest/oio_common/db_structures.py`,
but the template generation ignored it. This patch changes that and updates the
tests accordingly. Furthermore `titel` for klasse is also mandatory and has
been updated too.

As a nice side-benefit, we now have a reliable way to ask for a full list of
objects, because the postgres wildcard ("%") matches everything except null
(and there are no nulls now!).