Dont run all pg code as superuser
This change brings us back to where we once was... :p

The order of execution is quite a delicate thing and it also
matters a great deal who (or rather their permissions) executes
what code. When eveything is run as superuser subtle changes
occur that could be a headache down the line. For instance some
things were moved from actual_state (schema) to public.