Allow endpoints with and without trailing slashes.
This has been the source of a bit too much confussion :)

Werkzeug interprets url rules without a trailing slash to exactly that, but
we're not that strict. The ``strict_slashes`` has been set to ``False`` for
all endpoints, which means that they 301 without the trailing slash to their
slashed counter part.

I took the liberty to reformat the calls to ``Flask.add_url_rule``. I hope I
didn't step on anyones toes.