validation: we already know the request type, so don't extract it
The code currently fails when faced with requests containing two or
more attributes. Simply removing the logic for extracting the type is
easier than fixing it.

I removed the explicit tests for this functionality, and adjusted a
few tests that implicitly relied on this behaviour, by performing a
request targeting the wrong type. While at it, I added a test for
that.