Remove `mox` entry point from setup.py
Reverts 78be2e53bdf99ca5274a6497e46dd6418616a2fd and
ea1b7f883a9a7b63beca79d08db22b5d0afc460b

The mox entry point requires .egg-info to be updated at all times. An outdated
`.egg_info` can give you `DistributionNotFound` exception from `pkg_resources`
because `oio_rest.egg_info/requires.txt` and `requirements.txt` is not in sync.
This update can be done with the `python setup.pt egg_info` command. The problem
is that in docker, it is hard to do it automatically. We can run it in the on
image creation in Dockerfile as we did before this commit. It makes sure that
`.egg-info` is present inside the container with the right permissions.

However, if we mount `/code` to the host machine during development, we overwrite
the `.egg-info` generated during image creation. There are four possible ways to
solve this.

1) Generate a new `.egg-info` on container creation in docker-entrypoint.sh. The
problem with this approach is user permissions. When we want the `/code`
directory to be owned by root, but gunicorn to be run as another user. We could
properly solve this by using sudo or similar in docker-entrypoint.sh, but the
complexity will be even greater.

2) Let the user generate `.egg-info` on the host machine. This requires python
and dependencies on the host and is an extra opaque step in installing.

3) Don`t generate a `.egg-info`, but leave the `mox` executable in place. This
could cause hours of confusion when it works in some environments but not others.
It is especially devious as it points towards missing dependencies.

4) Remove the `mox` and use `python3 -m oio_rest` consistently.

This implements 4 as it is the simplest. KISS.