34486: Add requirements.in files, for better support for pinning versions
We no longer list the various requirements as dependencies in setup.py.
Reading and parsing the generated files (or the .in) files, is a hassle.
oio_rest was never installed in that way, without also installing the
related requirements files - which mimics the way it works for OS2mo

Ideally the requirements in the main .in file should be listed _only_
as requirements in setup.py, and we should let pip-compile generate
from that, but seeing as we still have a few non-PyPI dependencies,
getting that to work doesn't seem worth the effort at the moment.